backend/requirements_new.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pydantic
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
langchain-text-splitters>=0.0.0
python-dotenv>=1.0.0
